refactor: move deps, examples for as-needed setup
