+++ packages/agent-toolkit/eslint.config.js
@@ -1,4 +1,9 @@
 import { config } from "@recallnet/eslint-config/base";
 
-/** @type {import("eslint").Linter.Config} */
-export default config;
+/** @type {import("eslint").Linter.Config[]} */
+export default [
+  ...config,
+  {
+    ignores: ["examples/**"],
+  },
+];
