+++ packages/agent-toolkit/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@recallnet/agent-toolkit",
-  "version": "0.0.1",
+  "version": "0.0.0",
   "description": "Recall agent toolkit with framework agnostic adapters",
   "type": "module",
   "scripts": {
@@ -64,23 +64,42 @@
     }
   },
   "dependencies": {
-    "@ai-sdk/openai": "^1.2.5",
-    "@langchain/core": "^0.3.42",
-    "@langchain/openai": "^0.4.4",
-    "@modelcontextprotocol/sdk": "^1.7.0",
     "@recallnet/chains": "workspace:*",
     "@recallnet/sdk": "workspace:*",
-    "ai": "^4.1.61",
-    "dotenv": "^16.4.7",
-    "langchain": "^0.3.19",
-    "openai": "^4.87.3",
     "viem": "^2.22.9",
-    "zod": "^3.24.1",
+    "zod": "^3.24.1"
+  },
+  "peerDependencies": {
+    "@langchain/core": "^0.3.42",
+    "@modelcontextprotocol/sdk": "^1.7.0",
+    "ai": "^4.1.61",
+    "openai": "^4.87.3",
     "zod-to-json-schema": "^3.24.4"
   },
+  "peerDependenciesMeta": {
+    "@langchain/core": {
+      "optional": true
+    },
+    "@modelcontextprotocol/sdk": {
+      "optional": true
+    },
+    "ai": {
+      "optional": true
+    },
+    "openai": {
+      "optional": true
+    },
+    "zod-to-json-schema": {
+      "optional": true
+    }
+  },
   "devDependencies": {
+    "@ai-sdk/openai": "^1.2.5",
+    "@langchain/openai": "^0.4.4",
     "@recallnet/eslint-config": "workspace:*",
     "@recallnet/typescript-config": "workspace:*",
+    "dotenv": "^16.4.7",
+    "langchain": "^0.3.19",
     "tsup": "^8.3.6",
     "typescript": "^5.7.3"
   }
